fix nasty bug affecting fitting errors due to scope escape
The `perror` variable is used as a local buffer and then the address
is handed to the `mpfit` call.

As we then return the `mp_result` struct, the original `perror`
variable goes out of scope.

On `refc` this was never visible, as the GC would only zero the
`perror` memory on the next run (I assume?), which was unlikely right
after the `fit` call. But `arc/orc` being deterministic and `perror`
being `=destroy`ed at the end of the `fit` procedure, immediately
zeros the memory, resulting in "0" errors for every parameter. Neat!